requirements.txt
fastapi
uvicorn
jinja2
python-decouple
python-multipart
boto3
python-ffmpeg